+++ cockatrice/src/database/model/card/card_completer_proxy_model.h
