+++ cockatrice/src/database/model/card/card_search_model.cpp
@@ -1,6 +1,7 @@
 #include "card_search_model.h"
 
-#include "../utility/levenshtein.h"
+#include "../../../utility/levenshtein.h"
+#include "../database/card_database_model.h"
 
 #include <algorithm>
