+++ cockatrice/src/database/model/database/card_database_display_model.h
@@ -1,63 +1,14 @@
-#ifndef CARDDATABASEMODEL_H
-#define CARDDATABASEMODEL_H
+#ifndef COCKATRICE_CARD_DATABASE_DISPLAY_MODEL_H
+#define COCKATRICE_CARD_DATABASE_DISPLAY_MODEL_H
 
-#include "../filters/filter_string.h"
-#include "card_database.h"
+#include "../../../filters/filter_string.h"
 
-#include <QAbstractListModel>
 #include <QList>
 #include <QSet>
 #include <QSortFilterProxyModel>
 #include <QTimer>
 
 class FilterTree;
-
-class CardDatabaseModel : public QAbstractListModel
-{
-    Q_OBJECT
-public:
-    enum Columns
-    {
-        NameColumn,
-        SetListColumn,
-        ManaCostColumn,
-        PTColumn,
-        CardTypeColumn,
-        ColorColumn
-    };
-    enum Role
-    {
-        SortRole = Qt::UserRole
-    };
-    CardDatabaseModel(CardDatabase *_db, bool _showOnlyCardsFromEnabledSets, QObject *parent = nullptr);
-    ~CardDatabaseModel() override;
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override;
-    int columnCount(const QModelIndex &parent = QModelIndex()) const override;
-    QVariant data(const QModelIndex &index, int role) const override;
-    QVariant headerData(int section, Qt::Orientation orientation, int role = Qt::DisplayRole) const override;
-    CardDatabase *getDatabase() const
-    {
-        return db;
-    }
-    CardInfoPtr getCard(int index) const
-    {
-        return cardList[index];
-    }
-
-private:
-    QList<CardInfoPtr> cardList;
-    QSet<CardInfoPtr> cardListSet; // Supports faster lookups in cardDatabaseEnabledSetsChanged()
-    CardDatabase *db;
-    bool showOnlyCardsFromEnabledSets;
-
-    inline bool checkCardHasAtLeastOneEnabledSet(CardInfoPtr card);
-private slots:
-    void cardAdded(CardInfoPtr card);
-    void cardRemoved(CardInfoPtr card);
-    void cardInfoChanged(CardInfoPtr card);
-    void cardDatabaseEnabledSetsChanged();
-};
-
 class CardDatabaseDisplayModel : public QSortFilterProxyModel
 {
     Q_OBJECT
@@ -138,26 +89,4 @@ private slots:
     const QString sanitizeCardName(const QString &dirtyName, const QMap<wchar_t, wchar_t> &table);
 };
 
-class TokenDisplayModel : public CardDatabaseDisplayModel
-{
-    Q_OBJECT
-public:
-    explicit TokenDisplayModel(QObject *parent = nullptr);
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override;
-
-protected:
-    bool filterAcceptsRow(int sourceRow, const QModelIndex &sourceParent) const override;
-};
-
-class TokenEditModel : public CardDatabaseDisplayModel
-{
-    Q_OBJECT
-public:
-    explicit TokenEditModel(QObject *parent = nullptr);
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override;
-
-protected:
-    bool filterAcceptsRow(int sourceRow, const QModelIndex &sourceParent) const override;
-};
-
-#endif
+#endif // COCKATRICE_CARD_DATABASE_DISPLAY_MODEL_H
