+++ cockatrice/src/interface/widgets/visual_database_display/visual_database_display_widget.h
@@ -2,7 +2,7 @@
 #define VISUAL_DATABASE_DISPLAY_WIDGET_H
 
 #include "../../../database/card_database.h"
-#include "../../../database/card_database_model.h"
+#include "../../../database/model/database/card_database_model.h"
 #include "../../../deck/custom_line_edit.h"
 #include "../../../deck/deck_list_model.h"
 #include "../../../filters/filter_tree_model.h"
