+++ cockatrice/src/interface/widgets/visual_deck_editor/visual_deck_editor_widget.cpp
@@ -1,10 +1,10 @@
 #include "visual_deck_editor_widget.h"
 
-#include "../../../database/card_completer_proxy_model.h"
 #include "../../../database/card_database.h"
 #include "../../../database/card_database_manager.h"
-#include "../../../database/card_database_model.h"
-#include "../../../database/card_search_model.h"
+#include "../../../database/model/card/card_completer_proxy_model.h"
+#include "../../../database/model/card/card_search_model.h"
+#include "../../../database/model/database/card_database_model.h"
 #include "../../../deck/deck_list_model.h"
 #include "../../../deck/deck_loader.h"
 #include "../../../main.h"
