+++ cockatrice/src/tabs/abstract_tab_deck_editor.cpp
@@ -2,7 +2,7 @@
 
 #include "../client/tapped_out_interface.h"
 #include "../database/card_database_manager.h"
-#include "../database/card_database_model.h"
+#include "../database/model/database/card_database_model.h"
 #include "../deck/deck_stats_interface.h"
 #include "../dialogs/dlg_load_deck.h"
 #include "../dialogs/dlg_load_deck_from_clipboard.h"
