+++ cockatrice/src/tabs/visual_deck_storage/tab_deck_storage_visual.cpp
@@ -1,6 +1,6 @@
 #include "tab_deck_storage_visual.h"
 
-#include "../../database/card_database_model.h"
+#include "../../database/model/database/card_database_model.h"
 #include "../../interface/widgets/cards/deck_preview_card_picture_widget.h"
 #include "../../interface/widgets/visual_deck_storage/visual_deck_storage_widget.h"
 #include "../tab_supervisor.h"
